hooks/src/lib/useExpiryDate/useExpiryDate.ts
import { useState } from "react";

import useExpiryDateValidation from "./useExpiryDateValidation";

import { ExpiryDateKeys } from "../types/card-custom-hook";

const useExpiryDate = () => {
  const [expiryDate, setExpiryDate] = useState<Record<ExpiryDateKeys, string>>({
    month: "",
    year: "",
  });

  const { errorState, errorText, validateExpiryDate } = useExpiryDateValidation();

  const updateExpiryDate = (name: string, value: string) => {
    setExpiryDate((prevValue) => {
      return {
        ...prevValue,
        [name]: value,
      };
    });
  };

  const handleExpiryDateChange = (event: React.ChangeEvent<HTMLInputElement>) => {
    const { name, value } = event.target as {
      name: ExpiryDateKeys;
      value: string;
    };

    const canExpiryDateUpdate = validateExpiryDate(name, value, expiryDate);

    if (!canExpiryDateUpdate) return;

    updateExpiryDate(name, value);
  };

  const isExpiryDateCompleted =
    Object.values(expiryDate).every((cardNumber) => cardNumber.length === 2) && Object.values(errorState).every((isValid) => !isValid);

  return {
    expiryDate,
    errorState,
    isExpiryDateCompleted,
    errorText,
    handleExpiryDateChange,
  };
};

export default useExpiryDate;
